Move ReactionPath into separate file
